codegen/requirements.txt
case-converter==1.1.0
graphql-core==3.2.3